Update README.md
In two schools A and B, two different health programs are considered. In one school milk is distributed and in another school milk is not distributed. Nutrition experts plan to take these two classes by checking the height, weight and age of the students and compare them with each other. Write a program that takes the number of students in each class plus their age, height and weight information and stores each in a list and prints the average age, height and weight of each class in a separate line in the output (as float) and then print the class whose average height is higher in the output. (In the beginning, enter the number and information of class A, and then the information of class B. The output should be printed in the order of the input information.) If the average height is the same, the class with a lower average weight will be printed. If the average height and weight are equal, the expression Same (exactly the same) will be printed in the output